labs/official/mslearn-ai-information-extraction/Labfiles/05-rag-pipeline/requirements.txt
python-dotenv
azure-ai-contentunderstanding
azure-search-documents==11.6.0
openai
